Disable invalid O_CLOEXEC on PASE
AIX messed up O_CLOEXEC by setting it to a flag that's greater
than can be expressed as an integer. This was corrected in later
versions of AIX, but PASE 7.2 and 7.3 still have the incorrect
values. We work around this by adding bound checks to all uses
of O_CLOEXEC.