Run tests like AIX buildbots & add test reporting
Test arguments taken from various AIX buildbots at
https://buildbot.python.org/all/#/builders

The Python build suite has a --junit-xml option, but unlike the pytest
option of the same name, outputs in a different XML format which is not
understood by Jenkins.

Jenkins JUnit XML needs to comply with this XML Schema:
https://github.com/jenkinsci/xunit-plugin/blob/master/src/main/resources/org/jenkinsci/plugins/xunit/types/model/xsd/junit-10.xsd

While we could use the Jenkins xunit plugin with XSL to transform the
XML format, it's tricky enough that it's just easier to do in Python.

Co-authored-by: Abdirahim Musse <33973272+abmusse@users.noreply.github.com>