Work around PASE named semaphore deficiencies
PASE doesn't currently support inheriting named semaphores across a
fork. Instead, the semaphores become invalid and must be re-opened.

We work around this deficiency by re-opening the semaphore in
semlock_afterfork, similar to what is done in semlock_rebuild. In order
for this to work, though, the semaphore must remain linked in the
filesystem, so we disable calling sem_unlink in semlock_create. This
means that even when using the fork handler, the semlock will have a
name, causing it to be registered with the resource_tracker, which will
then call unlink it when it calls semlock_unlink.

This should not be a problem unless the parent ends, leaving the child
running and the child then forks a new process. In that child process
the afterfork handler would run, but the semaphore would have been
unlinked when its grandparent ended, thus preventing the grandchild from
re-opening the semaphore.